requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-cloud-aiplatform==1.108.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.17.0
google-cloud-storage==2.12.0
google-auth==2.23.4
google-api-core==2.14.0
redis==5.0.1
firebase-admin==6.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
PyJWT==2.8.0
requests==2.31.0
tenacity==8.2.3
ratelimit==2.2.1
